requirements.txt
Flask
Flask-JSONRPC
jsonrpcclient[requests]
newspaper3k
pika
pymongo
python-dateutil
python-jsonrpc
redis
requests
scikit-learn